Check Custom Banned Words Against Azure AD Banned Words Global & Per-Tenant List
With this PoSH script, you can test candidate banned words in a text file against Azure AD to determine if a word is already listed in either the global Microsoft list and/or the custom per-tenant list. As the custom per-tenant list has a current limit of 1000 you do not want to duplicate words

For more information, please see: Azure AD Password Protection (A.k.a. Banned Password List) - Optimizing The Custom Per Tenant List (Part 6)