Installing Azure AD Password Protection Proxy Service
With this PoSH script, one can install The Azure AD Password Protection Proxy Service on a Windows Server. Use the script, specify the servers to target and specify the full path to the installation source and you are good to go!

For more information, please see: (2019-10-22) Azure AD Password Protection (A.k.a. Banned Password List) – Deploy The Proxy Service (Part 3)