Installing Azure AD Password Protection DC Agent
With this PoSH script, one can install The Azure AD Password Protection DC Agent on a all RWDCs in the AD domain or AD forest. Use the script, specify the desired scope and specify the full path to the installation source and you are good to go!

For more information, please see: (2019-10-24) Azure AD Password Protection (A.k.a. Banned Password List) – Deploy The DC Agent (Part 4)